pyproject.toml
[tool.poetry]
name = "sxsw-hackathon-aibe"
version = "0.0.1"
description = "The AI BE for our app to kickstart your first months of employment and help track responsibilities - onboardsmart."
authors = ["Lukas Judith <lukasjudith12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
pyngrok = "^7.2.0"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
groq = "^0.11.0"
dspy = "^2.5.12"
dspy-ai = "^2.5.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
